RFC: adding a runtime ready RPI4 dtb for tests.
PleasePI4 DTBs as static file typically do not include memory stuff:
those are added by videocore firmware.

The DTB provided has been saved with U-Boot
just before transfering control to an  OS: this is the DTB that
fdt-rs will have to deal with in real world.

Another interest of the RPI4 DTB is that bus addresses need translations
using ranges to obtain MMIO region for devices.